Checkbox funcionando

+++ app/src/main/java/com/tute/estacionamientofrlp/CompraActivity.java
@@ -337,7 +337,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
 
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
 
@@ -354,7 +354,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
                     Log.e("ERROR3.2", String.valueOf(pateComp));
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
                     c6.setChecked(false);
@@ -376,7 +376,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
 
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
 
@@ -392,7 +392,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
                     Log.e("ERROR3.2", String.valueOf(pateComp));
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
                     c6.setChecked(false);
@@ -414,7 +414,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
 
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
 
@@ -431,7 +431,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
                     Log.e("ERROR3.2", String.valueOf(pateComp));
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
                     c6.setChecked(false);
@@ -454,7 +454,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
 
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
 
@@ -470,7 +470,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
                     Log.e("ERROR3.2", String.valueOf(pateComp));
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
                     c6.setChecked(false);
@@ -492,7 +492,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
 
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
 
@@ -509,7 +509,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
                     Log.e("ERROR3.2", String.valueOf(pateComp));
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Black));}
                     c6.setChecked(false);
